backend/requirements.txt
fastapi
uvicorn
transformers
requests
python-dotenv
pydantic